requirements.txt
bcrypt==4.0.1
build==1.0.3
fastapi==0.104.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
pydantic==2.4.2
pydantic_core==2.10.1
sendgrid==6.10.0
uvicorn==0.24.0
email-validator==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.3
argon2-cffi==23.1.0
